+++ blkshp_os/recipes/doctype/recipe_batch_ingredient/__init__.py
@@ -1,0 +1,5 @@
+"""
+Recipe Batch Ingredient child table package.
+"""
+
+